feat(layouts): consolidate "friends" and "files" layout into new "basic" layout
