+++ pages/files/browse/Browse.html
@@ -1,43 +1,48 @@
-<div
-  class="files-container"
-  :class="{'files-container-mobile': $device.isMobile}"
-  data-cy="files-screen"
->
-  <div class="files-top">
-    <FilesFilepath />
+<DroppableWrapper @handle-drop-prop="handleDrop">
+  <div
+    class="files-container"
+    :class="{'files-container-mobile': $device.isMobile}"
+    data-cy="files-screen"
+  >
+    <div class="files-top">
+      <FilesFilepath />
+    </div>
+    <FilesControls ref="controls" class="controls" />
+    <div class="files-aside-container">
+      <FilesAside />
+    </div>
+    <div class="files-view">
+      <TypographyText
+        v-if="!directory.length"
+        :text="$t('pages.files.empty')"
+      />
+      <FilesMobileList
+        v-else-if="$device.isMobile"
+        :directory="directory"
+        @like="like"
+        @share="share"
+        @rename="rename"
+        @remove="remove"
+        @handle="handle"
+      />
+      <FilesGrid
+        v-else-if="gridLayout"
+        :directory="directory"
+        @like="like"
+        @share="share"
+        @rename="rename"
+        @remove="remove"
+        @handle="handle"
+      />
+      <FilesList
+        v-else-if="!gridLayout"
+        :directory="directory"
+        @like="like"
+        @share="share"
+        @rename="rename"
+        @remove="remove"
+        @handle="handle"
+      />
+    </div>
   </div>
-  <FilesControls ref="controls" class="controls" />
-  <div class="files-aside-container">
-    <FilesAside />
-  </div>
-  <div class="files-view">
-    <TypographyText v-if="!directory.length" :text="$t('pages.files.empty')" />
-    <FilesMobileList
-      v-else-if="$device.isMobile"
-      :directory="directory"
-      @like="like"
-      @share="share"
-      @rename="rename"
-      @remove="remove"
-      @handle="handle"
-    />
-    <FilesGrid
-      v-else-if="gridLayout"
-      :directory="directory"
-      @like="like"
-      @share="share"
-      @rename="rename"
-      @remove="remove"
-      @handle="handle"
-    />
-    <FilesList
-      v-else-if="!gridLayout"
-      :directory="directory"
-      @like="like"
-      @share="share"
-      @rename="rename"
-      @remove="remove"
-      @handle="handle"
-    />
-  </div>
-</div>
+</DroppableWrapper>
