+++ pages/files/browse/index.vue
@@ -6,12 +6,16 @@ import { mapState, mapGetters } from 'vuex'
 import iridium from '~/libraries/Iridium/IridiumManager'
 import { IridiumItem } from '~/libraries/Iridium/files/types'
 import { RootState } from '~/types/store/store'
-import { ModalWindows } from '~/store/ui/types'
+import { ModalWindows, SettingsRoutes } from '~/store/ui/types'
 import { FileRouteEnum } from '~/libraries/Enums/enums'
+import DroppableWrapper from '~/components/ui/DroppableWrapper/DroppableWrapper.vue'
 
 export default Vue.extend({
   name: 'Files',
-  layout: 'files',
+  components: {
+    DroppableWrapper,
+  },
+  layout: 'basic',
   data() {
     return {
       items: iridium.files.state.items,
@@ -23,10 +27,13 @@ export default Vue.extend({
       path: (state) => (state as RootState).files.path,
       gridLayout: (state) => (state as RootState).files.gridLayout,
       modals: (state) => (state as RootState).ui.modals,
+      consentToScan: (state) =>
+        (state as RootState).textile.userThread.consentToScan,
     }),
     ...mapGetters({
       sortedItems: 'files/sortedItems',
     }),
+    ...mapGetters('textile', ['getInitialized']),
     directory(): IridiumItem[] {
       return this.sortedItems(this.items, this.$route.query.route)
     },
@@ -94,6 +101,41 @@ export default Vue.extend({
      */
     share(item: IridiumItem) {
       this.$toast.show(this.$t('todo - share') as string)
+    },
+    /**
+     * @method handleDrop
+     * @description Allows the drag and drop of files into the filesystem
+     * @param e Drop event data object
+     */
+    handleDrop(e: DragEvent) {
+      e.preventDefault()
+
+      if (!this.getInitialized) {
+        return
+      }
+
+      if (!this.consentToScan) {
+        this.$toast.error(
+          this.$t('pages.files.errors.enable_consent') as string,
+          {
+            duration: 3000,
+          },
+        )
+        this.$store.commit('ui/toggleSettings', {
+          show: true,
+          defaultRoute: SettingsRoutes.PRIVACY,
+        })
+        return
+      }
+
+      // if already uploading, return to prevent bucket fast-forward crash
+
+      if (e?.dataTransfer) {
+        const files: (File | null)[] = [...e.dataTransfer.items].map((f) =>
+          f.getAsFile(),
+        )
+        this.$refs.files?.$children[0].$refs.controls.handleFile(files)
+      }
     },
   },
 })
